pyproject.toml
[project]
name = "oai-agents"
version = "1.0.0"
description = "OpenAI Agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "motor>=3.7.0",
    "openai>=1.66.2",
    "openai-agents>=0.0.3",
    "pydantic-settings>=2.8.1",
]

[tool.uv]
dev-dependencies = ["mypy>=1.15.0", "ruff>=0.9.10"]

[tool.ruff]
exclude = [".git", ".mypy_cache", ".ruff_cache"]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I"]
fixable = ["ALL"]
mccabe = { max-complexity = 15 }

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.pytest_env]
CI = "true"

[tool.mypy]
plugins = ["pydantic.mypy"]

disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
disallow_untyped_decorators = false
ignore_missing_imports = true
implicit_reexport = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true
